Rest CRUD for Patients

+++ src/service/patient_service.py
@@ -27,7 +27,11 @@ def get_one_by_id(patient_id):
     :param patient_id:
     :return patient:
     """
-    return Patient.query.get_or_404(patient_id)
+    patient = Patient.query.get(patient_id)
+    if patient:
+        return patient
+    else:
+        raise RuntimeError(f"Patient with id: {patient_id} was not found")
 
 
 def update(patient_id, patient):
@@ -36,15 +40,18 @@ def update(patient_id, patient):
     :param patient_id:
     :param patient:
     """
-    patient_for_edit = Patient.query.get_or_404(patient_id)
-    patient_for_edit.full_name = patient.full_name
-    patient_for_edit.year_of_birth = patient.year_of_birth
-    patient_for_edit.kind_of_ache = patient.kind_of_ache
-    patient_for_edit.phone_number = patient.phone_number
-    patient_for_edit.email = patient.email
+    patient_for_edit = Patient.query.get(patient_id)
+    if patient_for_edit:
+        patient_for_edit.full_name = patient.full_name
+        patient_for_edit.year_of_birth = patient.year_of_birth
+        patient_for_edit.kind_of_ache = patient.kind_of_ache
+        patient_for_edit.phone_number = patient.phone_number
+        patient_for_edit.email = patient.email
 
-    db.session.add(patient_for_edit)
-    db.session.commit()
+        db.session.add(patient_for_edit)
+        db.session.commit()
+    else:
+        raise RuntimeError(f"Patient with id: {patient_id} was not found")
 
 
 def delete(patient_id):
@@ -52,6 +59,9 @@ def delete(patient_id):
     Function for deleting patient from DB
     :param patient_id:
     """
-    patient = Patient.query.get_or_404(patient_id)
-    db.session.delete(patient)
-    db.session.commit()
+    patient = Patient.query.get(patient_id)
+    if patient:
+        db.session.delete(patient)
+        db.session.commit()
+    else:
+        raise RuntimeError(f"Patient with id: {patient_id} was not found")
